code cleanup
1. secondy => secodary;
2. RT_DETACH_CPU => RT_CPU_DETACHED
3. rt_interrupt_check_schedule => rt_scheduler_do_irq_switch
