+++ rt-thread/bsp/qemu-vexpress-a9/cpu/start_gcc.S
@@ -355,9 +355,9 @@ vector_resv:
     bl      rt_hw_trap_resv
     b       .
 
-.global set_secondy_cpu_boot_address
-set_secondy_cpu_boot_address:
-    ldr r0, =secondy_cpu_start
+.global set_secondary_cpu_boot_address
+set_secondary_cpu_boot_address:
+    ldr r0, =secondary_cpu_start
 
     mvn r1, #0 //0xffffffff
     ldr r2, =0x10000034
@@ -365,8 +365,8 @@ set_secondy_cpu_boot_address:
     str r0, [r2, #-4]
     mov pc, lr
 
-.global secondy_cpu_start
-secondy_cpu_start:
+.global secondary_cpu_start
+secondary_cpu_start:
     mrc p15, 0, r1, c1, c0, 1
     mov r0, #(1<<6)
     orr r1, r0
@@ -390,7 +390,7 @@ secondy_cpu_start:
     cps #Mode_SVC
     ldr sp, =svc_stack_2_limit
 
-    b secondy_cpu_c_start
+    b secondary_cpu_c_start
 
 .bss
 .align 2   //align to  2~2=4
